% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quiz2functions.R
\name{accuracy}
\alias{accuracy}
\title{Compute accuracy based on prediction and reference truth value}
\usage{
accuracy(pred, truth)
}
\arguments{
\item{pred}{prediction result}

\item{truth}{reference(truth) result}
}
\value{
accuracy
}
\description{
Compute accuracy based on prediction and reference truth value
Data should be boolean, or 1 for positive and 0 for negative
}
\examples{
accuracy(c(1,0,1,1,0),c(1,0,1,1,1))
accuracy(c(T,F,T,T,F),c(F,T,T,F,F))
}
