% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homeworkfunctions.R
\name{adjR2}
\alias{adjR2}
\title{Compute Adjusted R square}
\usage{
adjR2(pred, truth, n)
}
\arguments{
\item{pred}{prediction result}

\item{truth}{reference(truth) result}

\item{n}{the number of predictors}
}
\value{
adjusted r square score
}
\description{
Compute Adjusted R square
}
\examples{
adjR2(1:5,2:6,1)
}
