% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homeworkfunctions.R
\name{bhAdjust}
\alias{bhAdjust}
\title{Bonferroni-Holm adjusted p-value}
\usage{
bhAdjust(p)
}
\arguments{
\item{p}{p-values}
}
\value{
boolean vector of each p-value is significant or not
}
\description{
Bonferroni-Holm adjusted p-value
}
\examples{
test.data = seq(0.0025,0.0250,0.0025)
test.shuffle = sample(test.data)
bhAdjust(test.data)
bhAdjust(test.shuffle)
}
