% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homeworkfunctions.R
\name{boxMuller}
\alias{boxMuller}
\title{Simulate Normal Data with Box Muller Tranformation}
\usage{
boxMuller(n = 1)
}
\arguments{
\item{n}{number of data to simulate}
}
\value{
simulated result, 1D vector with length n
}
\description{
Simulate Normal Data with Box Muller Tranformation
}
\examples{
sim = boxMuller(500)
}
