% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homeworkfunctions.R
\name{chiSquareCounts}
\alias{chiSquareCounts}
\title{chi-square test for count data in a tibble or similar object}
\usage{
chiSquareCounts(tib)
}
\arguments{
\item{tib}{count data in a tibble or similar object}
}
\value{
chisquare test result include chisq statistic, p-value, degree of freedom
}
\description{
chi-square test for count data in a tibble or similar object
}
\examples{
data = data.frame(x1=c(15,34), x2 =c(65,46))
chiSquareCounts(data)
}
