% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homeworkfunctions.R
\name{effectSize}
\alias{effectSize}
\title{effect size from data and group}
\usage{
effectSize(x, g)
}
\arguments{
\item{x}{numeric data}

\item{g}{group allocation of data}
}
\value{
effect size(cohen's d)
}
\description{
effect size from data and group
}
\examples{
effectSize(runif(20), sample(0:1,20, replace=T))
}
