% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quiz2functions.R
\name{f1}
\alias{f1}
\title{Compute f1 score based on prediction and reference truth value}
\usage{
f1(pred, truth)
}
\arguments{
\item{pred}{prediction result}

\item{truth}{reference(truth) result}
}
\value{
f1 score
}
\description{
Compute f1 score based on prediction and reference truth value
Data should be boolean, or 1 for positive and 0 for negative
}
\examples{
f1(c(1,0,1,1,0),c(1,0,1,1,1))
f1(c(T,F,T,T,F),c(F,T,T,F,F))
}
