% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homeworkfunctions.R
\name{fdrAdjust}
\alias{fdrAdjust}
\title{FDR adjusted p-value}
\usage{
fdrAdjust(p)
}
\arguments{
\item{p}{p-values}
}
\value{
boolean vector of each p-value is significant or not
}
\description{
FDR adjusted p-value
}
\examples{
test.data = seq(0.0025,0.0250,0.0025)
test.shuffle = sample(test.data)
fdrAdjust(test.data)
fdrAdjust(test.shuffle)
}
