% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homeworkfunctions.R
\name{minimumN}
\alias{minimumN}
\title{minimum sample size n with power=0.8}
\usage{
minimumN(d)
}
\arguments{
\item{d}{effect size (cohen's d)}
}
\value{
minimum sample size n
}
\description{
minimum sample size n with power=0.8
}
\examples{
minimumN(0.8)
}
