% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newfunctions.R
\name{pcApprox}
\alias{pcApprox}
\title{estimate the data with first npc pricipal components}
\usage{
pcApprox(x, npc)
}
\arguments{
\item{x}{original numeric data}

\item{npc}{number of pricipal components}
}
\value{
the estimation matrix
}
\description{
estimate the data with first npc pricipal components
}
\examples{
pcApprox(mtcars, 7)
}
