% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newfunctions.R
\name{pcLollipop}
\alias{pcLollipop}
\title{Create a lollipop plot of the principal component loadings of the (potentially unscaled/uncentered) data x}
\usage{
pcLollipop(x)
}
\arguments{
\item{x}{original numeric data}
}
\value{
the plot object
}
\description{
Create a lollipop plot of the principal component loadings of the (potentially unscaled/uncentered) data x
}
\examples{
pcLollipop(iris[,1:4])
pcLollipop(mtcars)
}
