% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homeworkfunctions.R
\name{postHocPower}
\alias{postHocPower}
\title{estimate of post-hoc power based on 1000 simultions}
\usage{
postHocPower(d, n1, n2)
}
\arguments{
\item{d}{effect size (cohen's d)}

\item{n1}{size of group 1}

\item{n2}{size of group 2}
}
\value{
estimated power
}
\description{
estimate of post-hoc power based on 1000 simultions
}
\examples{
postHocPower(1,12,13)
}
