% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quiz2functions.R
\name{ppv}
\alias{ppv}
\title{Compute precision based on prediction and reference truth value}
\usage{
ppv(pred, truth)
}
\arguments{
\item{pred}{prediction result}

\item{truth}{reference(truth) result}
}
\value{
precision
}
\description{
Compute precision based on prediction and reference truth value
Data should be boolean, or 1 for positive and 0 for negative
}
\examples{
ppv(c(1,0,1,1,0),c(1,0,1,1,1))
ppv(c(T,F,T,T,F),c(F,T,T,F,F))
}
