% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homeworkfunctions.R
\name{r2}
\alias{r2}
\title{Compute R square}
\usage{
r2(pred, truth)
}
\arguments{
\item{pred}{prediction result}

\item{truth}{reference(truth) result}
}
\value{
r square score
}
\description{
Compute R square
}
\examples{
r2(1:5,2:6)
}
