% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quiz2functions.R
\name{sensitivity}
\alias{sensitivity}
\title{Compute Sensitivity based on prediction and reference truth value}
\usage{
sensitivity(pred, truth)
}
\arguments{
\item{pred}{prediction result}

\item{truth}{reference(truth) result}
}
\value{
sensitivity
}
\description{
Compute Sensitivity based on binary prediction and reference truth value.
Data should be boolean, or 1 for positive and 0 for negative
}
\examples{
sensitivity(c(1,0,1,1,0),c(1,0,1,1,1))
sensitivity(c(T,F,T,T,F),c(F,T,T,F,F))
}
