% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quiz2functions.R
\name{specificity}
\alias{specificity}
\title{Compute specificity based on prediction and reference truth value}
\usage{
specificity(pred, truth)
}
\arguments{
\item{pred}{prediction result}

\item{truth}{reference(truth) result}
}
\value{
specificity
}
\description{
Compute specificity based on prediction and reference truth value
Data should be boolean, or 1 for positive and 0 for negative
}
\examples{
specificity(c(1,0,1,1,0),c(1,0,1,1,1))
specificity(c(T,F,T,T,F),c(F,T,T,F,F))
}
