% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homeworkfunctions.R
\name{twoSidedT}
\alias{twoSidedT}
\title{two sided t test}
\usage{
twoSidedT(t, df)
}
\arguments{
\item{t}{t-statistic}

\item{df}{degree of freedom}
}
\value{
p-value as decimal
}
\description{
two sided t test
}
\examples{
twoSidedT(2,5)
}
