% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homeworkfunctions.R
\name{twoSidedZ}
\alias{twoSidedZ}
\title{two sided Z test}
\usage{
twoSidedZ(z)
}
\arguments{
\item{z}{z-statistic}
}
\value{
p-value as decimal
}
\description{
two sided Z test
}
\examples{
twoSidedZ(2)
}
