% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newfunctions.R
\name{unscale}
\alias{unscale}
\title{unscale the object return by scale function}
\usage{
unscale(x)
}
\arguments{
\item{x}{scaled object}
}
\value{
the un-scaled matrix
}
\description{
unscale the object return by scale function
}
\examples{
mtcars |> scale() |> unscale()
}
