% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homeworkfunctions.R
\name{welchT}
\alias{welchT}
\title{welch t test}
\usage{
welchT(x, y)
}
\arguments{
\item{x}{dataset1}

\item{y}{dataset2 to compare}
}
\value{
class Welch.t.test. includes p.value, t-statistic, degree of freedom
mean of dataset1, mean of dataset2
}
\description{
welch t test
}
\examples{
welchT(c(1:10), c(7:20))
}
